backend/requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==2.0.12
cryptography==43.0.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2
django-extensions==3.2.3
django-filter==23.1
django-import-export==3.3.9
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
et-xmlfile==1.1.0
flake8==7.1.0
idna==3.7
iniconfig==2.0.0
MarkupPy==1.14
mccabe==0.7.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
pluggy==0.13.1
psycopg2-binary==2.9.9
py==1.11.0
pycodestyle==2.12.0
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
pytest==6.2.4
pytest-django==4.4.0
pytest-lazy-fixture==0.6.3
pytest-pythonpath==0.7.3
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.26.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tablib==3.5.0
toml==0.10.2
typing_extensions==4.12.2
urllib3==1.26.19
webcolors==1.11.1
xlrd==2.0.1
xlwt==1.3.0
